requirements.txt
# ============================================================
# Market Intelligence Terminal — Python Dependencies
# ============================================================
# Install:  pip install -r requirements.txt
# ============================================================

# ── Data Pipeline ───────────────────────────────────────────
yfinance>=0.2.40
pandas>=2.2.2
numpy>=1.26.4
httpx>=0.27.0

# ── Database ────────────────────────────────────────────────
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.30
alembic>=1.13.1

# ── API Server ──────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
pydantic>=2.7.4
pydantic-settings>=2.3.4

# ── AI Engine ───────────────────────────────────────────────
ollama>=0.2.1

# ── Analysis ────────────────────────────────────────────────
networkx>=3.3

# ── Scheduling ──────────────────────────────────────────────
apscheduler>=3.10.4

# ── Utilities ───────────────────────────────────────────────
python-dotenv>=1.0.1
loguru>=0.7.2
tqdm>=4.66.4
